ennamoo

+++ ft_strlcpy.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcpy.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: liovino <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: liovino <liovino@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/12 14:43:54 by liovino           #+#    #+#             */
-/*   Updated: 2024/12/12 14:43:55 by liovino          ###   ########.fr       */
+/*   Updated: 2024/12/15 12:27:41 by liovino          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (src[i] != '\0' && i < size)
